requirements.txt
babel==2.17.0
blinker==1.9.0
cffi==2.0.0
click==8.1.8
cryptography==46.0.3
Flask==3.1.1
flask-babel==4.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.2.3
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
-e git+ssh://git@github.com/kdevy/myflaskbase.git@825c95c07e831f11b65246d136c118e0b5bd4410#egg=myflaskbase
pycparser==2.23
PyMySQL==1.1.1
python-dotenv==1.1.1
pytz==2025.2
SQLAlchemy==2.0.41
typing_extensions==4.14.1
Werkzeug==3.1.3
WTForms==3.2.1
zipp==3.23.0
